Remove use of git-lfs for handling database
git-lfs was difficult to work with due to the no public forks issue:
https://medium.com/@megastep/github-s-large-file-storage-is-no-panacea-for-open-source-quite-the-opposite-12c0e16a9a91

This removes the intermediate database from version control. For the moment
the user will need to generate it themselves (which in most cases would be the
case outside of the lab anyway). For in-house purposes I've put a copy of the T-drive.