move route data function to the core script
This function can be used in most of the other data scripts, so I'm moving it out of the elev one. Also hardcoding the crs because that should never change for the BBS lat/longs. It will be adjusted by the extract() functions if a data source crs is different.